% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb.R
\name{as.data.frame.ladybug.query_result.QueryResult}
\alias{as.data.frame.ladybug.query_result.QueryResult}
\title{Convert a Ladybug Query Result to a Data Frame}
\usage{
\method{as.data.frame}{ladybug.query_result.QueryResult}(x, ...)
}
\arguments{
\item{x}{A Ladybug query result object.}

\item{...}{Additional arguments passed to \code{as.data.frame}.}
}
\value{
An R \code{data.frame} containing the query results.
}
\description{
Provides an S3 method to seamlessly convert a Ladybug query result object into a
standard R \code{data.frame}.
}
\examples{
\donttest{
conn <- lb_connection(":memory:")
lb_execute(conn, "CREATE NODE TABLE User(name STRING, age INT64,
PRIMARY KEY (name))")
lb_execute(conn, "CREATE (:User {name: 'Alice', age: 25})")
result <- lb_execute(conn, "MATCH (a:User) RETURN a.name, a.age")

# Convert the result to a data.frame
df <- as.data.frame(result)
print(df)
}
}
