% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_graph.R
\name{as_igraph}
\alias{as_igraph}
\title{Convert a Ladybug Query Result to an igraph Object}
\usage{
as_igraph(query_result)
}
\arguments{
\item{query_result}{A \code{ladybug_query_result} object from \code{lb_execute()} that
contains a graph.}
}
\value{
An \code{igraph} object.
}
\description{
Converts a Ladybug query result into an \code{igraph} graph object.
}
\details{
This function takes a \code{ladybug_query_result} object and extracts nodes and edges
directly from the query results, then constructs an \code{igraph} object. It is
the final step in the \code{lb_execute -> as_igraph} workflow.
}
\examples{
\dontrun{
if (requireNamespace("igraph", quietly = TRUE)) {
  conn <- lb_connection(":memory:")
  lb_execute(conn, "CREATE NODE TABLE Person(name STRING,
  PRIMARY KEY (name))")
  lb_execute(conn, "CREATE REL TABLE Knows(FROM Person TO Person)")
  lb_execute(conn, "CREATE (p:Person {name: 'Alice'}),
  (q:Person {name: 'Bob'})")
  lb_execute(conn, "MATCH (a:Person), (b:Person) WHERE
                                                    a.name='Alice' AND
                                                    b.name='Bob'
                                                    CREATE (a)-[:Knows]->(b)"
)

  res <- lb_execute(conn, "MATCH (p:Person)-[k:Knows]->(q:Person)
  RETURN p, k, q")
  g <- as_igraph(res)
  print(g)
  rm(conn, res, g)
}
}
}
