% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb.R
\name{as_tibble.ladybug.query_result.QueryResult}
\alias{as_tibble.ladybug.query_result.QueryResult}
\title{Convert a Ladybug Query Result to a Tibble}
\usage{
\method{as_tibble}{ladybug.query_result.QueryResult}(x, ...)
}
\arguments{
\item{x}{A Ladybug query result object.}

\item{...}{Additional arguments passed to \code{as_tibble}.}
}
\value{
A \code{tibble} containing the query results.
}
\description{
Provides an S3 method to convert a Ladybug query result object into a
\code{tibble}. This requires the \code{tibble} package to be installed.
}
\examples{
\dontrun{
if (requireNamespace("tibble", quietly = TRUE)) {
  conn <- lb_connection(":memory:")
  lb_execute(conn, "CREATE NODE TABLE User(name STRING, age INT64,
  PRIMARY KEY (name))")
  lb_execute(conn, "CREATE (:User {name: 'Alice', age: 25})")
  result <- lb_execute(conn, "MATCH (a:User) RETURN a.name, a.age")

  # Convert the result to a tibble
  tbl <- tibble::as_tibble(result)
  print(tbl)
}
}
}
