% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_graph.R
\name{as_tidygraph}
\alias{as_tidygraph}
\title{Convert a Ladybug Query Result to a tidygraph Object}
\usage{
as_tidygraph(query_result)
}
\arguments{
\item{query_result}{A \code{ladybug_query_result} object from \code{lb_execute()} that
contains a graph.}
}
\value{
A \code{tbl_graph} object.
}
\description{
Converts a Ladybug query result into a \code{tidygraph} \code{tbl_graph} object.
}
\examples{
\donttest{
if (requireNamespace("tidygraph", quietly = TRUE)) {
  conn <- lb_connection(":memory:")
  lb_execute(conn, "CREATE NODE TABLE Person(name STRING,
  PRIMARY KEY (name))")
  lb_execute(conn, "CREATE (p:Person {name: 'Alice'})")
  res <- lb_execute(conn, "MATCH (p:Person) RETURN p")
  g_tidy <- as_tidygraph(res)
  print(g_tidy)
  rm(conn, res, g_tidy)
}
}
}
