% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb.R
\name{lb_get_column_data_types}
\alias{lb_get_column_data_types}
\title{Get Column Data Types from a Query Result}
\usage{
lb_get_column_data_types(result)
}
\arguments{
\item{result}{A Ladybug query result object.}
}
\value{
A character vector of column data types.
}
\description{
Retrieves the data types of the columns in a Ladybug query result.
}
\examples{
\dontrun{
conn <- lb_connection(":memory:")
lb_execute(conn, "CREATE NODE TABLE User(name STRING, age INT64,
PRIMARY KEY (name))")
lb_execute(conn, "CREATE (:User {name: 'Alice', age: 25})")
result <- lb_execute(conn, "MATCH (a:User) RETURN a.name, a.age")
lb_get_column_data_types(result)
}
}
