% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_load_data.R
\name{lb_merge_df}
\alias{lb_merge_df}
\title{Merge Data from a Data Frame into Ladybug using a Merge Query}
\usage{
lb_merge_df(conn, df, merge_query)
}
\arguments{
\item{conn}{A Ladybug connection object.}

\item{df}{A \code{data.frame} or \code{tibble} containing the data to merge.}

\item{merge_query}{A string representing the Ladybug query for merging data.}
}
\value{
This function is called for its side effect of merging data and does
not return a value.
}
\description{
This function is intended for merging data from an R \code{data.frame} into Ladybug
using a specified merge query.
}
\examples{
\dontrun{
conn <- lb_connection(":memory:")

lb_execute(conn, "CREATE NODE TABLE Person(name STRING, current_city STRING,
PRIMARY KEY (name))")
lb_execute(conn, "CREATE NODE TABLE Item(name STRING, PRIMARY KEY (name))")
lb_execute(conn, "CREATE REL TABLE PURCHASED(FROM Person TO Item)")

my_data <- data.frame(
  name = c("Alice", "Bob"),
  item = c("Book", "Pen"),
  current_city = c("New York", "London")
)

merge_statement <- "
MERGE (p:Person {name: df.name})
MERGE (i:Item {name: df.item})
MERGE (p)-[:PURCHASED]->(i)
ON MATCH SET p.current_city = df.current_city
ON CREATE SET p.current_city = df.current_city
"

lb_merge_df(conn, my_data, merge_statement)

result <- lb_execute(conn, "MATCH (p:Person)-[:PURCHASED]->(i:Item)
RETURN p.name, i.name, p.current_city")
print(as.data.frame(result))
}
}
\seealso{
\href{https://ladybugdb.com/import/copy-from-dataframe}{Ladybug Copy from DataFrame}
}
